cleaning up messages
